MBS-11738 Bind DeployStep.kt to uploadCrashlyticsMappingFile
