+++ subprojects/gradle/cd/src/main/kotlin/com/avito/ci/steps/deploy/UploadCrashlyticsProguardFileTasksProvider.kt
@@ -1,0 +1,35 @@
+package com.avito.ci.steps.deploy
+
+import com.avito.cd.BuildVariant
+import com.avito.cd.CdBuildConfig
+import org.gradle.api.Project
+import java.util.Locale
+
+internal class UploadCrashlyticsProguardFileTasksProvider {
+
+    fun provide(
+        project: Project,
+        deployments: List<CdBuildConfig.Deployment.GooglePlay>
+    ): List<String> = deployments
+        .groupBy { it.buildVariant }
+        .keys
+        .map { buildVariant ->
+            uploadCrashlyticsProguardFileTaskProvider(project, buildVariant)
+        }
+
+    private fun uploadCrashlyticsProguardFileTaskProvider(
+        project: Project,
+        buildVariant: BuildVariant
+    ): String {
+        val projectName = project.name
+        val buildVariantCapitalized =
+            buildVariant.toString()
+                .toLowerCase(Locale.getDefault())
+                .capitalize(Locale.getDefault())
+        /**
+         * TODO when I try to find task provider via project.tasks.named I get Exception that there is no task
+         * That's because firebase-crashlytics-plugin creates task some how after we trying bind to it
+         */
+        return ":$projectName:uploadCrashlyticsMappingFile$buildVariantCapitalized"
+    }
+}
